fix: add System import for MathF and disambiguate Color constructors
- Add 'using System' so MathF.Sin resolves
- Cast Color args to byte to resolve overload ambiguity

+++ raylibcs/Program.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Runtime.InteropServices.JavaScript;
 using Raylib_cs;
 
@@ -20,13 +21,13 @@ public partial class Application
         byte alpha = (byte)(180 + (int)(75 * MathF.Sin(timer * 2f)));
 
         Raylib.BeginDrawing();
-        Raylib.ClearBackground(new Color(30, 30, 46, 255));
+        Raylib.ClearBackground(new Color((byte)30, (byte)30, (byte)46, (byte)255));
 
         Raylib.DrawFPS(10, 10);
 
         Raylib.DrawText("RaylibCS + WASM", 200, 180, 40, Color.Green);
-        Raylib.DrawText("Powered by Raylib-cs 7.0 & .NET 8", 180, 240, 20, new Color(200, 200, 200, alpha));
-        Raylib.DrawText("Running in your browser!", 250, 290, 20, new Color(130, 130, 230, 255));
+        Raylib.DrawText("Powered by Raylib-cs 7.0 & .NET 8", 180, 240, 20, new Color((byte)200, (byte)200, (byte)200, alpha));
+        Raylib.DrawText("Running in your browser!", 250, 290, 20, new Color((byte)130, (byte)130, (byte)230, (byte)255));
 
         Raylib.EndDrawing();
     }
